Move `h.schemas.admin_group` -> `h.schemas.forms.admin.group`
